@@ -1,6 +1,6 @@
 package com.melkonian.example.lesson3;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public interface FragmentNavigator {
   void startSecondFragment(@NonNull String shape);
