@@ -1,9 +1,10 @@
 package com.melkonian.example.lesson3.activities;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import com.melkonian.example.lesson3.FragmentNavigator;
 import com.melkonian.example.lesson3.R;
 
@@ -11,10 +12,14 @@ public class MainActivity extends AppCompatActivity implements FragmentNavigator
   private MainFragment mainFragment;
   private SecondFragment secondFragment;
 
+  private FragmentManager fragmentManager;
+
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_main);
+
+    fragmentManager = getSupportFragmentManager();
 
     mainFragment = new MainFragment();
     secondFragment = new SecondFragment();
@@ -29,12 +34,15 @@ public class MainActivity extends AppCompatActivity implements FragmentNavigator
   @Override
   public void startSecondFragment(@NonNull String shape) {
     secondFragment.setShape(shape);
-    FragmentManager fragmentManager = getSupportFragmentManager();
+    commitOperation(secondFragment);
+
+  }
+
+  private void commitOperation(@NonNull Fragment fragmentToCommit) {
     fragmentManager
         .beginTransaction()
-        .replace(R.id.fragment_container, secondFragment)
+        .replace(R.id.fragment_container, fragmentToCommit)
         .addToBackStack("")
         .commit();
-
   }
 }
